+++ tests/test_cli/test_validation.py
@@ -4,12 +4,6 @@ import pytest
 from typer import BadParameter
 
 from matcha_ml.cli._validation import (
-    LONGEST_RESOURCE_NAME,
-    MAXIMUM_RESOURCE_NAME_LEN,
-    _check_length,
-    _is_alphanumeric,
-    _is_not_digits,
-    _is_valid_prefix,
     find_closest_matches,
     get_command_validation,
     prefix_typer_callback,
@@ -75,91 +69,6 @@ def test_region_typer_callback_invalid():
     """Test that the callback raises a BadParameter error when invalid input is given."""
     with pytest.raises(BadParameter):
         region_typer_callback("ukwset")
-
-
-def test_is_valid_prefix_expected():
-    """Test that the is valid prefix functions behaves as expected with correct input."""
-    assert _is_valid_prefix("matcha") == "matcha"
-
-
-@pytest.mark.parametrize(
-    "prefix, expectation", [("mat-cha", False), ("matcha", True), ("matcha123", True)]
-)
-def test_is_alphanumeric(prefix: str, expectation: bool):
-    """Test that the _is_alphanumeric internal function works as expected.
-
-    Args:
-        prefix (str): the test prefix.
-        expectation (bool): what the expected result is.
-    """
-    assert _is_alphanumeric(prefix) == expectation
-
-
-@pytest.mark.parametrize(
-    "prefix, expectation", [("thisisareallylongprefix", False), ("matcha", True)]
-)
-def test_check_length(prefix: str, expectation: bool):
-    """Test that the _check_length internal function works as expected.
-
-    Args:
-        prefix (str): the test prefix.
-        expectation (bool): what the expected result is.
-    """
-    assert _check_length(prefix) == expectation
-
-
-@pytest.mark.parametrize(
-    "prefix, expectation", [("1234", False), ("matcha", True), ("matcha123", True)]
-)
-def test_is_not_digits(prefix: str, expectation: bool):
-    """Test that _is_not_digits works as expected.
-
-    Args:
-        prefix (str): the test prefix.
-        expectation (bool): what the expected result is.
-    """
-    assert _is_not_digits(prefix) == expectation
-
-
-@pytest.mark.parametrize(
-    "prefix, error_msg, expectation",
-    [
-        (
-            "1234",
-            "Resource group name prefix cannot contain only numbers.",
-            MatchaInputError,
-        ),
-        (
-            "matcha&&",
-            "Resource group name prefix can only contain alphanumeric characters.",
-            MatchaInputError,
-        ),
-        (
-            "mat-cha",
-            "Resource group name prefix can only contain alphanumeric characters.",
-            MatchaInputError,
-        ),
-        (
-            "thisisareallylongprefixmatcha",
-            f"Resource group name prefix must be between 3 and {MAXIMUM_RESOURCE_NAME_LEN - len(LONGEST_RESOURCE_NAME)} characters long.",
-            MatchaInputError,
-        ),
-    ],
-)
-def test_is_valid_prefix_invalid(
-    prefix: str, error_msg: str, expectation: MatchaInputError
-):
-    """Test that the is_valid_prefix function behaves as expected when invalid input is given.
-
-    Args:
-        prefix (str): the invalid prefix.
-        error_msg (str): the specific error message that is expected.
-        expectation (MatchaInputError): the error that is expected to be raised.
-    """
-    with pytest.raises(expectation) as err:
-        _is_valid_prefix(prefix)
-
-    assert str(err.value) == error_msg
 
 
 @pytest.mark.parametrize(
